reestruturar como os gcodes customizados funcionam agora eles são interpretados como gcodes quaisquer vantagens: 	- podemos reusar a interpretação de gcode por serial que vem junto 	do marlin 	- não precisamos dar parse no gcode 2x na hora de lidar com os 'L' 	- facilita imensamente a implementação de envio de gcode do app desvantagens: 	- complica levemente a implementação de gcodes que expandem para 	vários gcodes, como o 'L3'

+++ Marlin/lucas/Estacao.h
@@ -53,7 +53,7 @@ public:
 
     void atualizar_status(std::string_view str) const;
 
-    int posicao_absoluta() const;
+    float posicao_absoluta() const;
 
     size_t numero() const;
